limits calculator added
