--- funwithphysics/src/Components/IntegrationAndDifferentiation/Topics/Calculator.css
@@ -1,197 +1,0 @@
-.Calculator__main {
-    margin: auto;
-    padding: 3em;
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    grid-template-areas:
-      "header header"
-      "details formula"
-      "details process"
-      "example example"
-      "calc calc";
-    column-gap: 3em;
-    row-gap: 1.5em;
-    max-width: 100%;
-  }
-  .K_energy {
-    background-color: white !important;
-  }
-  
-  .Calculator__header {
-    grid-area: header;
-    align-self: center;
-    text-align: center;
-    font-size: 1.8em;
-    color: #e84545;
-    font-weight: 900;
-  }
-  
-  .Calculator__details {
-    grid-area: details;
-    align-self: center;
-    text-align: center;
-    font-size: 1.3em;
-    font-weight: bold;
-    justify-self: center;
-    word-wrap: break-word;
-    width: 30em;
-  }
-  
-  .Calculator__formula {
-    grid-area: formula;
-    font-family: math;
-  }
-  .Calculator__example {
-    grid-area: example;
-    align-self: start;
-    text-align: left;
-    margin-left: 20px;
-  }
-  
-  .Calculator__calc {
-    grid-area: calc;
-  }
-  
-  .Calculator__calc hr {
-    width: auto;
-    border: 1px dashed red;
-  }
-  
-  .Calculator__calc h3 {
-    font-size: 1.5em;
-    font-weight: bold;
-    color: #e84545;
-    text-align: center;
-  }
-  
-  .Calculator__process {
-    grid-area: process;
-    word-wrap: break-word;
-    width: 37.5em;
-    font-size: 1.2rem;
-    padding-right: 2vh;
-  }
-  
-  .straight_calc {
-    background: #f4efe3;
-    padding: 1rem;
-    width: 100%;
-  }
-  
-  .group {
-    display: flex;
-    box-sizing: border-box;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  .coordinate {
-    display: flex;
-    width: 45%;
-    padding-right: 5px;
-    margin-bottom: 5px;
-  }
-  .group label {
-    padding-left: 4px;
-    display: block;
-  }
-  .formula-table {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-    box-shadow: 0 0.1rem 1rem rgba(0, 0, 0, 0.1);
-  }
-  .formulae {
-    table-layout: fixed;
-    height: auto;
-    width: 90%;
-  }
-  /* Column widths are based on these cells */
-  .row-ID {
-    width: 50%;
-  }
-  .row-name {
-    width: 50%;
-  }
-  
-  .formulae td {
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-  .formulae th {
-    background: lightblue;
-    color: #24252a;
-  }
-  .formulae td,
-  .formulae th {
-    text-align: left;
-    padding: 5px 10px;
-  }
-  .formulae tr:nth-child(even) {
-    background: lightblue;
-  }
-  .formulae section h3 {
-    white-space: nowrap;
-  }
-  @media only screen and (max-width: 992px) {
-    .Calculator__main {
-      grid-template-columns: 1fr;
-      grid-template-areas: "header header" "details details" "formula formula " "process process " "example example" "calc calc";
-      row-gap: 3em;
-    }
-  
-    .Calculator__header,
-    .Calculator__details,
-    .Calculator__formula,
-    .Calculator__calc h3,
-    .Calculator__process {
-      text-align: center;
-      max-width: 100%;
-    }
-  
-    .Calculator__example {
-      text-align: center;
-      margin: 0;
-  
-    }
-  
-    
-    .Calculator__formula h3 {
-      margin-left: -20px;
-    }
-  }
-  
-  @media only screen and (max-width: 668px) {
-    .Calculator__header,
-    .Calculator__details,
-    .Calculator__formula,
-    .Calculator__calc h3,
-    .Calculator__process {
-      max-width: 100%;
-    }
-    .Calculator__main {
-      padding: 1.5rem;
-    }
-    .Calculator__calc hr {
-      max-width: 100%;
-    }
-    .Calculator__details,
-    .Calculator__process p {
-      text-align: justify;
-    }
-    .Calculator__process {
-      padding: 1rem;
-    }
-    input {
-      width: 100%;
-    }
-    .Calculator__calc hr {
-      width: 100%;
-      margin: 0px auto 0px auto;
-    }
-    .Calculator__calc form {
-      padding: 1rem;
-    }
-  }
-
